# Speedup bytecode-gen by applying an upper bound on `read_exact`