Tested with [the `amm` repo](https://github.com/sway-libs/amm):
Previously:
```console
❯ forc b --time-phases --silent
...
  Time elapsed to compile to ast: 4.9242855s
  Time elapsed to generate JSON ABI program: 49.625µs
  Time elapsed to compile ast to asm: 51.795523958s
  Time elapsed to compile asm to bytecode: 27.532206458s
```
With this change:
```console
❯ forc b --time-phases --silent
...
  Time elapsed to compile to ast: 4.970219875s
  Time elapsed to generate JSON ABI program: 48.375µs
  Time elapsed to compile ast to asm: 52.703999375s
  Time elapsed to compile asm to bytecode: 76.786833ms
```
Specifically looking at the last step which is ~360X faster.