@otrho 2022-10-21
This might actually alleviate the need for the `NOOP` blob that I'm putting into by far jumps branch, since the time wasted there was in this phase.